Moving repository logic to domain module
