+++ app/src/main/kotlin/com/arturogutierrez/openticator/di/module/ApplicationModule.kt
@@ -4,14 +4,11 @@ import android.content.Context
 import android.content.SharedPreferences
 import android.view.LayoutInflater
 import com.arturogutierrez.openticator.domain.account.repository.AccountRepository
-import com.arturogutierrez.openticator.domain.account.repository.AccountRepositoryImpl
 import com.arturogutierrez.openticator.domain.category.CategorySelector
 import com.arturogutierrez.openticator.domain.category.repository.CategoryRepository
-import com.arturogutierrez.openticator.domain.category.repository.CategoryRepositoryImpl
 import com.arturogutierrez.openticator.domain.cloud.FirebaseSessionManager
 import com.arturogutierrez.openticator.domain.cloud.SessionManager
 import com.arturogutierrez.openticator.domain.issuer.repository.IssuerRepository
-import com.arturogutierrez.openticator.domain.issuer.repository.IssuerRepositoryImpl
 import com.arturogutierrez.openticator.executor.JobExecutor
 import com.arturogutierrez.openticator.executor.MainThread
 import com.arturogutierrez.openticator.executor.PostExecutionThread
@@ -58,15 +55,15 @@ class ApplicationModule(val context: Context) {
 
   @Singleton
   @Provides
-  fun provideAccountRepository(accountDiskDataStore: AccountDiskDataStore): AccountRepository = AccountRepositoryImpl(accountDiskDataStore)
+  fun provideAccountRepository(accountDiskDataStore: AccountDiskDataStore) = AccountRepository(accountDiskDataStore)
 
   @Singleton
   @Provides
-  fun provideCategoryRepository(categoryDiskDataStore: CategoryDiskDataStore): CategoryRepository = CategoryRepositoryImpl(categoryDiskDataStore)
+  fun provideCategoryRepository(categoryDiskDataStore: CategoryDiskDataStore) = CategoryRepository(categoryDiskDataStore)
 
   @Singleton
   @Provides
-  fun provideIssuerRepository(issuerRepository: IssuerRepositoryImpl): IssuerRepository = issuerRepository
+  fun provideIssuerRepository() = IssuerRepository()
 
   @Singleton
   @Provides
