+++ data/src/test/kotlin/com/arturogutierrez/openticator/domain/account/repository/AccountRepositoryTest.kt
@@ -19,7 +19,7 @@ class AccountRepositoryTest {
 
   @Before
   fun setUp() {
-    accountRepository = AccountRepositoryImpl(mockAccountDataStore)
+    accountRepository = AccountRepository(mockAccountDataStore)
   }
 
   @Test
