+++ domain/src/main/kotlin/com/arturogutierrez/openticator/domain/issuer/repository/IssuerRepository.kt
@@ -3,8 +3,14 @@ package com.arturogutierrez.openticator.domain.issuer.repository
 import com.arturogutierrez.openticator.domain.issuer.model.Issuer
 
 import rx.Observable
+import javax.inject.Inject
+import javax.inject.Singleton
 
-interface IssuerRepository {
+class IssuerRepository @Inject constructor() {
 
   val issuers: Observable<List<Issuer>>
+    get() {
+      val allIssuers = Issuer.values().asList().filter { it != Issuer.UNKNOWN }
+      return Observable.just<List<Issuer>>(allIssuers)
+    }
 }
